Various corner-case fixes for possible numerical overflow, undefined behavior, and buffer sizes.

+++ drvrmem.c
@@ -616,7 +616,7 @@ int mem_compress_open(char *filename, int rwmode, int *hdl)
 	modulosize  = buffer[0];
 	modulosize |= buffer[1] << 8;
 	modulosize |= buffer[2] << 16;
-	modulosize |= buffer[3] << 24;
+	modulosize |= ((unsigned int)buffer[3]) << 24;
 
 /*
   the field ISIZE in the gzipped file header only stores 4 bytes and contains
@@ -664,7 +664,7 @@ int mem_compress_open(char *filename, int rwmode, int *hdl)
 	modulosize  = buffer[0];
 	modulosize |= buffer[1] << 8;
 	modulosize |= buffer[2] << 16;
-	modulosize |= buffer[3] << 24;
+	modulosize |= ((unsigned int)buffer[3]) << 24;
         finalsize = modulosize;
 
         estimated = 0;  /* file size is known, not estimated */
